fixes after review
