+++ packages/shared/src/components/inputs/index.ts
@@ -4,7 +4,6 @@ export { default as AmountInput } from './AmountInput.svelte'
 export { default as Checkbox } from './Checkbox.svelte'
 export { default as ClosableInput } from './ClosableInput.svelte'
 export { default as Dropdown } from './Dropdown.svelte'
-export { default as Dropdown2 } from './Dropdown2.svelte'
 export { default as Dropzone } from './Dropzone.svelte'
 export { default as Input } from './Input.svelte'
 export { default as InputContainer } from './InputContainer.svelte'
